+++ client/src/components/layout/RestrictedRoute.tsx
@@ -1,6 +1,6 @@
 import { Navigate, Outlet } from "react-router-dom";
 
-import { useAuthContext } from "~/hooks/useAuthContext";
+import { useAuthContext } from "~/hooks";
 import { ROUTE } from "~/utils/constants";
 
 interface ProtectedRouteProps {
